ui: code gardened the views/cluster file structure
