+++ packages/app/src/client/views/routes.tsx
@@ -27,9 +27,9 @@ import SelectedModule from "./module/SelectedModule";
 import Modules from "./module/Modules";
 import ChatView from "./chat";
 import HistoryView from "./history";
-import UserDetail from "./cluster/UserDetail";
-import TenantDetail from "./cluster/TenantDetail";
-import AlertmanagerConfigEditor from "./cluster/AlertmanagerConfigEditor";
+import UserDetail from "client/views/user/UserDetail";
+import TenantDetail from "client/views/tenant/TenantDetail";
+import AlertmanagerConfigEditor from "client/views/tenant/alertmanager/AlertmanagerConfigEditor";
 import HelpDialog from "./help";
 import NotFound from "./404/404";
 import FullPage from "client/layout/FullPage";
