+++ packages/app/src/client/views/user/UserPicker.tsx
@@ -15,45 +15,45 @@
  */
 
 import React from "react";
-
-import { PickerOption, usePickerService } from "client/services/Picker";
-import { useCommandService } from "client/services/Command";
 import { useHistory } from "react-router-dom";
-import useTenantList from "state/tenant/hooks/useTenantList";
-import { Tenant } from "state/tenant/types";
 
-function tenantToPickerOption(tenant: Tenant): PickerOption {
-  return {
-    text: tenant.name,
-    id: tenant.name
-  };
-}
+import { useCommandService } from "client/services/Command";
 
-const TenantPicker = () => {
+import { User } from "state/user/types";
+
+import useUserList from "state/user/hooks/useUserList";
+import { PickerOption, usePickerService } from "client/services/Picker";
+
+export const userToPickerOption = (user: User): PickerOption => ({
+  text: user.email,
+  id: user.id
+});
+
+const UserPicker = () => {
   const history = useHistory();
-  const tenants = useTenantList();
+  const users = useUserList();
 
   const { activatePickerWithText } = usePickerService(
     {
-      activationPrefix: "tenant:",
-      options: tenants ? tenants.map(tenantToPickerOption) : [],
+      activationPrefix: "user:",
+      options: users ? users.map(userToPickerOption) : [],
       onSelected: option => {
-        history.push(`/cluster/tenants/${option.id}`);
+        history.push(`/cluster/users/${option.id}`);
       }
     },
-    [tenants, history]
+    [users, history]
   );
 
   useCommandService({
-    id: "select-tenant-picker",
-    description: "Select Tenant",
+    id: "select-user-picker",
+    description: "Select User",
     handler: e => {
       e.keyboardEvent?.preventDefault();
-      activatePickerWithText("tenant: ");
+      activatePickerWithText("user: ");
     }
   });
 
   return null;
 };
 
-export default React.memo(TenantPicker);
+export default React.memo(UserPicker);
